Extract tag-and-release script. Update README
