+++ src/node-index.ts
@@ -3,4 +3,4 @@ export * from './logger';
 export * from './processing';
 export * from './config-parsing';
 export * from './config-hash';
-export * from './scripts';
+export * from './release-scripts';
